-DealHands method does now also deal community-cards if properly called (is not yet called for comcard-rounds though)
-began implementation of bettingRound. Big and Smallblinds are being
paid. Players can technically bet, however the bet is not yet connected
to neither the pot nor the players chipstack. Also, im in the middle of
this so there is not yet a loop in bettingRound, meaning there is just
one player to once fold/raise or call for testing-purposes